[WIP] Bundle compiler works with components
But not {{ember-components}} or {{component helper}} yet.
