+++ packages/@glimmer/runtime/test/ember-component-new-test.ts
@@ -1,113 +1,3 @@
-import { module, AbstractRenderTest, test, classes } from "@glimmer/test-helpers";
-
-class EmberishComponentTests extends AbstractRenderTest {
-  @test
-  "non-block without properties"() {
-    this.render({
-      layout: 'In layout'
-    });
-
-     this.assertComponent('In layout');
-     this.assertStableRerender();
-  }
-
-  @test
-  "block without properties"() {
-    this.render({
-      layout: 'In layout -- {{yield}}',
-      template: 'In template'
-    });
-
-    this.assertComponent('In layout -- In template');
-    this.assertStableRerender();
-  }
-
-  @test
-  "yield inside a conditional on the component"() {
-    this.render({
-      layout: 'In layout -- {{#if @predicate}}{{yield}}{{/if}}',
-      template: 'In template',
-      args: { predicate: 'predicate' },
-    }, { predicate: true });
-
-    this.assertComponent('In layout -- In template', {});
-    this.assertStableRerender();
-
-    this.rerender({predicate: false });
-    this.assertComponent('In layout -- <!---->');
-    this.assertStableNodes();
-
-    this.rerender({ predicate: true });
-    this.assertComponent('In layout -- In template', {});
-    this.assertStableNodes();
-  }
-
-  @test
-  "non-block with properties on attrs"() {
-    this.render({
-      layout: 'In layout - someProp: {{@someProp}}',
-      args: { someProp: '"something here"'}
-    });
-
-    this.assertComponent('In layout - someProp: something here');
-    this.assertStableRerender();
-  }
-
-  @test
-  "block with properties on attrs"() {
-    this.render({
-      layout: 'In layout - someProp: {{@someProp}} - {{yield}}',
-      template: 'In template',
-      args: { someProp: '"something here"' }
-    });
-
-    this.assertComponent('In layout - someProp: something here - In template');
-    this.assertStableRerender();
-  }
-
-  @test({ skip: true, kind: 'curly' })
-  "with ariaRole specified"() {
-    this.render({
-      layout: 'Here!',
-      attributes: { id: '"aria-test"', ariaRole: '"main"' }
-    });
-
-    this.assertComponent('Here!', { id: '"aria-test"', role: '"main"' });
-    this.assertStableRerender();
-  }
-
-  @test({ skip: true, kind: 'curly' })
-  "with ariaRole and class specified"() {
-    this.render({
-      layout: 'Here!',
-      attributes: { id: '"aria-test"', class: '"foo"', ariaRole: '"main"' }
-    });
-
-    this.assertComponent('Here!', { id: '"aria-test"', class: classes('ember-view foo'), role: '"main"' });
-    this.assertStableRerender();
-  }
-
-  @test({ skip: true, kind: 'curly' })
-  "with ariaRole specified as an outer binding"() {
-    this.render({
-      layout: 'Here!',
-      attributes: { id: '"aria-test"', class: '"foo"', ariaRole: 'ariaRole' }
-    }, { ariaRole: 'main' });
-
-    this.assertComponent('Here!', { id: '"aria-test"', class: classes('ember-view foo'), role: '"main"' });
-    this.assertStableRerender();
-  }
-
-  @test({ skip: true, kind: 'glimmer' })
-  "glimmer component with role specified as an outer binding and copied"() {
-    this.render({
-      layout: 'Here!',
-      attributes: { id: '"aria-test"', role: 'myRole' }
-    }, { myRole: 'main' });
-
-    this.assertComponent('Here!', { id: '"aria-test"', role: '"main"' });
-    this.assertStableRerender();
-  }
-}
+import { module, EmberishComponentTests } from "@glimmer/test-helpers";
 
 module('Emberish Components', EmberishComponentTests, { componentModule: true });
