+++ packages/@glimmer/test-helpers/lib/environment/components.ts
@@ -1,0 +1,4 @@
+export * from './components/basic';
+export * from './components/emberish-curly';
+export * from './components/emberish-glimmer';
+export * from './components/tagless';
