+++ packages/@glimmer/test-helpers/lib/environment/helper.ts
@@ -1,0 +1,46 @@
+
+import { Opaque, Dict } from "@glimmer/interfaces";
+import { PathReference, Tag, VOLATILE_TAG, Reference } from "@glimmer/reference";
+import { CapturedArguments, Arguments } from "@glimmer/runtime";
+
+export type UserHelper = (args: ReadonlyArray<Opaque>, named: Dict<Opaque>) => Opaque;
+
+export class HelperReference implements PathReference<Opaque> {
+  private helper: UserHelper;
+  private args: CapturedArguments;
+  public tag: Tag = VOLATILE_TAG;
+
+  constructor(helper: UserHelper, args: Arguments) {
+    this.helper = helper;
+    this.args = args.capture();
+  }
+
+  value() {
+    let { helper, args } = this;
+
+    return helper(args.positional.value(), args.named.value());
+  }
+
+  get(prop: string): SimplePathReference<Opaque> {
+    return new SimplePathReference(this, prop);
+  }
+}
+
+export class SimplePathReference<T> implements PathReference<T> {
+  private parent: Reference<T>;
+  private property: string;
+  public tag: Tag = VOLATILE_TAG;
+
+  constructor(parent: Reference<T>, property: string) {
+    this.parent = parent;
+    this.property = property;
+  }
+
+  value(): T {
+    return this.parent.value()[this.property];
+  }
+
+  get(prop: string): PathReference<Opaque> {
+    return new SimplePathReference(this, prop);
+  }
+}
