+++ packages/@glimmer/test-helpers/lib/environment/iterable.ts
@@ -1,0 +1,141 @@
+
+import { OpaqueIterator, IterationItem, AbstractIterable, Tag, Reference } from "@glimmer/reference";
+import { Opaque, Option } from "@glimmer/interfaces";
+import { UpdatableReference } from "@glimmer/object-reference";
+
+export type KeyFor<T> = (item: Opaque, index: T) => string;
+
+export class Iterable implements AbstractIterable<Opaque, Opaque, IterationItem<Opaque, Opaque>, UpdatableReference<Opaque>, UpdatableReference<Opaque>> {
+  public tag: Tag;
+  private ref: Reference<Opaque>;
+  private keyFor: KeyFor<Opaque>;
+
+  constructor(ref: Reference<Opaque>, keyFor: KeyFor<Opaque>) {
+    this.tag = ref.tag;
+    this.ref = ref;
+    this.keyFor = keyFor;
+  }
+
+  iterate(): OpaqueIterator {
+    let { ref, keyFor } = this;
+
+    let iterable = ref.value() as any;
+
+    if (Array.isArray(iterable)) {
+      return iterable.length > 0 ? new ArrayIterator(iterable, keyFor) : EMPTY_ITERATOR;
+    } else if (iterable === undefined || iterable === null) {
+      return EMPTY_ITERATOR;
+    } else if (iterable.forEach !== undefined) {
+      let array: Opaque[] = [];
+      for (let i = 0; i < iterable.length; i++) {
+        array.push(iterable[i]);
+      }
+      return array.length > 0 ? new ArrayIterator(array, keyFor) : EMPTY_ITERATOR;
+    } else if (typeof iterable === 'object') {
+      let keys = Object.keys(iterable);
+
+      if (keys.length > 0) {
+        let values: Opaque[] = [];
+
+        for (let i = 0; i < keys.length; i++) {
+          let key = keys[i];
+          values[i] = iterable[key];
+        }
+
+        return new ObjectKeysIterator(keys, values, keyFor);
+      } else {
+        return EMPTY_ITERATOR;
+      }
+    } else {
+      throw new Error(`Don't know how to {{#each ${iterable}}}`);
+    }
+  }
+
+  valueReferenceFor(item: IterationItem<Opaque, Opaque>): UpdatableReference<Opaque> {
+    return new UpdatableReference(item.value);
+  }
+
+  updateValueReference(reference: UpdatableReference<Opaque>, item: IterationItem<Opaque, Opaque>) {
+    reference.update(item.value);
+  }
+
+  memoReferenceFor(item: IterationItem<Opaque, Opaque>): UpdatableReference<Opaque> {
+    return new UpdatableReference(item.memo);
+  }
+
+  updateMemoReference(reference: UpdatableReference<Opaque>, item: IterationItem<Opaque, Opaque>) {
+    reference.update(item.memo);
+  }
+}
+
+class EmptyIterator implements OpaqueIterator {
+  isEmpty(): boolean {
+    return true;
+  }
+
+  next(): IterationItem<Opaque, Opaque> {
+    throw new Error(`Cannot call next() on an empty iterator`);
+  }
+}
+
+class ObjectKeysIterator implements OpaqueIterator {
+  private keys: string[];
+  private values: Opaque[];
+  private keyFor: KeyFor<string>;
+  private position = 0;
+
+  constructor(keys: string[], values: Opaque[], keyFor: KeyFor<string>) {
+    this.keys = keys;
+    this.values = values;
+    this.keyFor = keyFor;
+  }
+
+  isEmpty(): boolean {
+    return this.keys.length === 0;
+  }
+
+  next(): Option<IterationItem<Opaque, string>> {
+    let { position, keys, values, keyFor } = this;
+
+    if (position >= keys.length) return null;
+
+    let value = values[position];
+    let memo = keys[position];
+    let key = keyFor(value, memo);
+
+    this.position++;
+
+    return { key, value, memo };
+  }
+}
+
+class ArrayIterator implements OpaqueIterator {
+  private array: Opaque[];
+  private keyFor: KeyFor<number>;
+  private position = 0;
+
+  constructor(array: Opaque[], keyFor: KeyFor<number>) {
+    this.array = array;
+    this.keyFor = keyFor;
+  }
+
+  isEmpty(): boolean {
+    return this.array.length === 0;
+  }
+
+  next(): Option<IterationItem<Opaque, number>> {
+    let { position, array, keyFor } = this;
+
+    if (position >= array.length) return null;
+
+    let value = array[position];
+    let key = keyFor(value, position);
+    let memo = position;
+
+    this.position++;
+
+    return { key, value, memo };
+  }
+}
+
+const EMPTY_ITERATOR = new EmptyIterator();
