+++ packages/@glimmer/test-helpers/lib/suites.ts
@@ -1,2 +1,3 @@
 export * from './suites/initial-render';
 export * from './suites/components';
+export * from './suites/ember-components';
